nodejs

+++ updater.sh
@@ -294,7 +294,8 @@ Installazione di FFMpeg
 	    apt-cyg install openssl
 	fi
 
-	if ! command -v node &>/dev/null
+	if ! command -v node &>/dev/null &&
+		[ ! -f $SHARE/node.exe ]
 	then
 	    print_c 1 "Installazione di Nodejs.exe in $SHARE"
 	    wget -O $SHARE/node.exe https://nodejs.org/dist/v4.4.4/win-x86/node.exe
